code to generate a list of probable IPs (and /24s)
The algorithm is currently hard-coded to be the same as the approach used in the trinocular paper (or at least as close as I can make it by reverse-engineering what the paper says).

There still seems to be some discrepancy between the numbers quoted in the paper and what this tool generates. Will investigate more.