Revamp adaptive and recovery algorithm; add debug output
This makes the probing at once closer to, and further from, the algorithm described in the paper.
Now we require the belief to stabilize before probing stops (closer to the paper). This has the potential to cause more /24s to flip-flop state.
We also trigger recovery probing as soon as adaptive probing is done if the state is transitioning from UP to something else (not explicitly described in the paper). This acts as a last-ditch attempt to get the /24 back to UP.