Drop metas for /24s with >255
As the inline comment notes, we have a couple of /24s that have >255 pieces of metadata attached to them which is causing this assert to fire. Since there are only two (currently), we're just going to drop the extra meta information. We'll have to revisit this ~hack~ fix if/when we have many /24s with this behavior.